fetch backfills logs in GQL (#22038)
## Summary & Motivation
Fetches backfill logs using the captured log manager. 

Adds functionality to the captured log manager to read a sequence of
files in a directory as if they are a single file by sorting the file
names and then reading through the files N lines at a time. The majority
of the logic is in the base CapturedLogManager class, but each
implementor class will need to implement a method to get the log keys
for a given directory. I implemented this method for the
LocalComputeLogManager to enable testing, and follow up PRs can
implement it for the remaining ComputeLogManagers. internal pr
https://github.com/dagster-io/internal/pull/9879 that implements the
method for the cloud compute log manager

Fetching logs via GQL is gated on an instance method we can override to
control roll out

## How I Tested These Changes
new unit tests for the pagination scheme, added a unit test where we run
a backfill and then fetch the logs